04-Generative-AI/Deeplearning.ai/Fast-Prototyping-of-GenAI-Apps-with-Streamlit/M1/Lesson_3/requirements.txt
streamlit
groq
python-dotenv
matplotlib
pandas
altair
plotly